Update bvDHT.py
organized existing functions and added some comments. I also mapped out some finger table functionality. We needed a way to determine who was currently connected to the DHT in order to add them to our finger table as fingers. I mapped out a way to send and receive this information.